fixed eslint issues in the types folder and moved InputBoxPropsTypes
