+++ types/components/FigurePropsType.ts
@@ -1,4 +1,4 @@
 export enum Float {
   LEFT = 'left',
-  RIGHT = 'right',
-};
+  RIGHT = 'right'
+}
