+++ types/components/community/CardPropsType.ts
@@ -1,4 +1,4 @@
 export enum CardType {
   SMALL = 'small',
-  LARGE = 'large',
+  LARGE = 'large'
 }
